Refactor module structure by renaming files and updating import paths for data preparation, image filtering, image processing, mask processing, and splitter modules
