+++ data_preparation/image_processing/image_loading.py
