+++ data_preparation/mask_processing/buildings_masks.py
