+++ data_preparation/mask_processing/roads_masks.py
